Local Files and Launcher support
This just adds a new int function that's basically the main function, just with a removed apply_environment function call. This also edits the apply_environment function by changing the throw error into calling the new localfile function so instead of it tossing a error it just launches the old launcher that was exe dependent and used the local game files. This means you can now take the file from within the xlabs appdata folder and use it in the game files. I would also recommend allowing the exe be downloadable within the github page so people that cant use the launcher due to unknown problems can still play the game, just with a alternative link. I am recommending this as putting this into the official branch will allow the game to still be updated with the official release.